TCPMv2: Add host command VDM:Attention retrieval
Add onto the existing VDM_RESPONSE host command to allow the AP to
retrieve VDM:Attention messages.

BRANCH=None
BUG=b:208884535
TEST=build, deploy on skyrim and confirm ectool shows Attention messages
from the queue until it empties

+++ include/usb_pd_dpm_sm.h
@@ -118,8 +118,8 @@ void dpm_consume_attention(int port, uint8_t vdo_objects, uint32_t *buf);
  * @param[out] buf		Buffer to copy VDM header and VDO
  * @param[out] items_left	Number of Attention messages left in the queue
  */
-void dpm_vdm_attention_pop(int port, int *length, uint32_t *buf,
-			   int *items_left);
+void dpm_vdm_attention_pop(int port, uint8_t *length, uint32_t *buf,
+			   uint8_t *items_left);
 
 /*
  * Determines the current allocation for the connection, past the basic
